Move witness_version from ScriptExt to Script
The witness_version function is used in various is_* functions and must
be present in primitives. This can be moved to primitives wholesale and
removed from the extension trait.

Move witness_version from ScriptExt to Script.
